review changes - try catch null handling, minor refactor
