+++ bin/accessibility-automation/plugin/index.js
@@ -6,30 +6,26 @@ const browserstackAccessibility = (on, config) => {
   on('before:browser:launch', (browser = {}, launchOptions) => {
     try {
 
-      if (browser.name !== 'chrome') {
-        console.log(`Accessibility Automation will run only on Chrome browsers.`);
-        browser_validation = false;
+      if (process.env.ACCESSIBILITY_EXTENSION_PATH !== undefined) {
+        if (browser.name !== 'chrome') {
+          console.log(`Accessibility Automation will run only on Chrome browsers.`);
+          browser_validation = false;
+        }
+        if (browser.name === 'chrome' && browser.majorVersion <= 94) {
+          console.log(`Accessibility Automation will run only on Chrome browser version greater than 94.`);
+          browser_validation = false;
+        }
+        if (browser.isHeadless === true) {
+          console.log(`Accessibility Automation will not run on legacy headless mode. Switch to new headless mode or avoid using headless mode.`);
+          browser_validation = false;
+        }
+        if (browser_validation) {
+          const ally_path = path.dirname(process.env.ACCESSIBILITY_EXTENSION_PATH)
+          launchOptions.extensions.push(ally_path);
+          return launchOptions
+        }
       }
-      if (browser.majorVersion <= 94) {
-        console.log(`Accessibility Automation will run only on Chrome browser version greater than 94.`);
-        browser_validation = false;
-      }
-      if (browser.isHeadless === true) {
-        console.log(`Accessibility Automation will not run on legacy headless mode. Switch to new headless mode or avoid using headless mode.`);
-        browser_validation = false;
-      }
-
-      if (process.env.ACCESSIBILITY_EXTENSION_PATH === undefined) {
-        browser_validation = false
-        return
-      }
-
-      if (browser_validation) {
-        const ally_path = path.dirname(process.env.ACCESSIBILITY_EXTENSION_PATH)
-        launchOptions.extensions.push(ally_path);
-        return launchOptions
-      }
-    } catch {}
+    } catch(err) {}
     
   })
   config.env.ACCESSIBILITY_EXTENSION_PATH = process.env.ACCESSIBILITY_EXTENSION_PATH
